fix wrapping code

+++ life/src/life.c
@@ -6,7 +6,7 @@
 #include <time.h>
 
 #define WRAP(v, i, o, d) \
-    v = i + o, v = v < 0 ? d - 1 : v % d
+    v = (long) i + o, v = v < 0 ? d - 1 : v % d
 
 const unsigned N = 500;
